Waveform starts when user presses play, and stops after track is finished playing
