+++ Main/JamChat/JamChat/Base.lproj/Main.storyboard
@@ -87,21 +87,20 @@
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" fixedFrame="YES" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="97" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="LDL-Vh-qmE">
                                 <rect key="frame" x="0.0" y="0.0" width="375" height="437"/>
                                 <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                                <view key="tableHeaderView" contentMode="scaleToFill" id="XZL-Zu-yVb">
-                                    <rect key="frame" x="0.0" y="64" width="375" height="69"/>
-                                    <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                                </view>
                                 <prototypes>
                                     <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="blue" hidesAccessoryWhenEditing="NO" indentationLevel="1" indentationWidth="0.0" reuseIdentifier="MessageCell" rowHeight="97" id="Ret-99-1AO" customClass="MessageCell" customModule="JamChat" customModuleProvider="target">
-                                        <rect key="frame" x="0.0" y="161" width="375" height="97"/>
+                                        <rect key="frame" x="0.0" y="92" width="375" height="97"/>
                                         <autoresizingMask key="autoresizingMask"/>
                                         <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="Ret-99-1AO" id="AEP-NG-LYZ">
                                             <rect key="frame" x="0.0" y="0.0" width="375" height="96.5"/>
                                             <autoresizingMask key="autoresizingMask"/>
                                             <subviews>
+                                                <view contentMode="scaleToFill" fixedFrame="YES" translatesAutoresizingMaskIntoConstraints="NO" id="XZL-Zu-yVb">
+                                                    <rect key="frame" x="0.0" y="0.0" width="375" height="58"/>
+                                                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                                                </view>
                                                 <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="wgN-HT-JQh">
-                                                    <rect key="frame" x="172" y="34" width="30" height="30"/>
+                                                    <rect key="frame" x="165" y="66" width="44" height="30"/>
                                                     <state key="normal" title="Play"/>
                                                     <connections>
                                                         <action selector="onPlay:" destination="Ret-99-1AO" eventType="touchUpInside" id="v9N-lp-C9P"/>
@@ -109,6 +108,9 @@
                                                 </button>
                                             </subviews>
                                         </tableViewCellContentView>
+                                        <connections>
+                                            <outlet property="waveformView" destination="XZL-Zu-yVb" id="l7h-nr-ITR"/>
+                                        </connections>
                                     </tableViewCell>
                                 </prototypes>
                             </tableView>
@@ -117,7 +119,6 @@
                     </view>
                     <connections>
                         <outlet property="tableView" destination="LDL-Vh-qmE" id="rYZ-sJ-onP"/>
-                        <outlet property="waveformView" destination="XZL-Zu-yVb" id="voU-kK-jN7"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="VsT-5S-Ncc" userLabel="First Responder" sceneMemberID="firstResponder"/>
